:sparkles: 관리자 페이지 항목 복구
